requirements.txt
astroid==1.6.5
backports.functools-lru-cache==1.5
Click==7.0
configparser==3.7.3
enum34==1.1.6
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
isort==4.3.15
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==1.9.4
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.3.1
Werkzeug==0.14.1
wrapt==1.11.1
Authlib==0.14.3
oauth==1.0.1
oauthlib==3.1.0
requests==2.21.0
Flask-MySQLdb==0.2.0
Flask-Mail==0.9.1
image==1.5.33
Flask-MySQL==1.5.2
flask-mysql-connector==1.1.0